ft: moved dashboard links into site.ts
